+++ pytsite/content/__init__.py
@@ -3,8 +3,7 @@
 # Public API
 from . import _model as model, _widget as widget
 from ._api import register_model, get_models, find, get_model, get_model_title, dispense, get_sections, \
-    create_section, create_tag, get_tags, get_tag, get_statuses, get_section, is_model_registered, \
-    generate_rss
+    dispense_section, get_tags, dispense_tag, get_statuses, is_model_registered, generate_rss
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
@@ -14,7 +13,6 @@ __license__ = 'MIT'
 def __init():
     """Module Init Wrapper.
     """
-    import sys
     from pytsite import admin, taxonomy, settings, console, assetman, odm, events, tpl, lang, router, robots, browser
     from . import _eh
     from ._model import Tag, Section, ContentSubscriber