+++ pytsite/content_import/_error.py
@@ -1,0 +1,13 @@
+"""PytSite Content Import Errors.
+"""
+__author__ = 'Alexander Shepetko'
+__email__ = 'a@shepetko.com'
+__license__ = 'MIT'
+
+
+class DriverNotRegistered(Exception):
+    pass
+
+
+class ImportError(Exception):
+    pass